Restructure self-tests to apply them archunit-junit6* modules, too
- Required because the previous setting was unable to test junit5-engine and junit6-engine at the same time
