+++ archunit-example/example-junit6/build.gradle
@@ -1,5 +1,6 @@
 plugins {
     id 'archunit.java-examples-conventions'
+    id 'archunit.self-test-conventions'
 }
 
 ext.moduleName = 'com.tngtech.archunit.example.junit6'
@@ -8,6 +9,11 @@ ext.minimumJavaVersion = JavaVersion.VERSION_17
 dependencies {
     testImplementation project(path: ':archunit-junit6')
     testImplementation project(path: ':archunit-example:example-plain')
+
+    archTestImplementation project(path: ':archunit-junit6')
+    archTestRuntimeOnly libs.junit6PlatformLauncher
+    // We want to run the architecture self-tests on the `test` sourceSet of this subproject
+    archTestRuntimeOnly sourceSets.test.output
 }
 
 test {
