+++ archunit-example/example-plain/src/archTest/java/com/tngtech/archunit/exampletest/ArchUnitExamplePlainArchitectureTest.java
@@ -1,0 +1,13 @@
+package com.tngtech.archunit.exampletest;
+
+import com.tngtech.archunit.ArchUnitExampleArchitectureRules;
+import com.tngtech.archunit.junit.AnalyzeClasses;
+import com.tngtech.archunit.junit.ArchTest;
+import com.tngtech.archunit.junit.ArchTests;
+
+@AnalyzeClasses(packages = {"com.tngtech.archunit.example", "com.tngtech.archunit.exampletest"})
+class ArchUnitExamplePlainArchitectureTest {
+
+    @ArchTest
+    static final ArchTests example_rules = ArchTests.in(ArchUnitExampleArchitectureRules.class);
+}
