+++ archunit-junit/junit4/src/archTest/resources/archunit.properties
@@ -1,0 +1,1 @@
+archRule.failOnEmptyShould=false
