+++ archunit-junit/junit5/api/src/archTest/resources/archunit.properties
@@ -1,0 +1,1 @@
+archRule.failOnEmptyShould=false
