+++ archunit-junit/junit5/engine-api/build.gradle
@@ -1,6 +1,7 @@
 plugins {
     id 'archunit.java-release-conventions'
     id 'archunit.java-junit-conventions'
+    id 'archunit.self-test-conventions'
 }
 
 ext.moduleName = 'com.tngtech.archunit.junit5.engineapi'
@@ -13,15 +14,13 @@ dependencies {
 
     testImplementation project(path: ':archunit-junit5-api')
     testImplementation libs.assertj
+
+    // Important: Use the version of archunit-junit* that matches this package.
+    archTestImplementation project(path: ':archunit-junit5')
+    archTestRuntimeOnly libs.junit5PlatformLauncher
 }
 
 compileJava.dependsOn project(':archunit-junit5-api').finishArchive
-
-test {
-    useJUnitPlatform() {
-        excludeEngines 'archunit'
-    }
-}
 
 shadowJar {
     exclude 'META-INF/maven/**'
