+++ archunit-junit/junit5/engine/build.gradle
@@ -1,6 +1,7 @@
 plugins {
     id 'archunit.java-release-conventions'
     id 'archunit.java-junit-conventions'
+    id 'archunit.self-test-conventions'
 }
 
 ext.moduleName = 'com.tngtech.archunit.junit5.engine'
@@ -20,6 +21,10 @@ dependencies {
     testImplementation libs.mockito
     testImplementation libs.mockito.junit5
     testImplementation libs.log4j.core
+
+    // Important: Use the version of archunit-junit* that matches this package.
+    archTestImplementation project(path: ':archunit-junit5')
+    archTestRuntimeOnly libs.junit5PlatformLauncher
 }
 
 gradle.projectsEvaluated {
