+++ archunit-junit/junit6/engine/build.gradle
@@ -1,6 +1,7 @@
 plugins {
     id 'archunit.java-release-conventions'
     id 'archunit.java-junit-conventions'
+    id 'archunit.self-test-conventions'
 }
 
 ext.moduleName = 'com.tngtech.archunit.junit6.engine'
@@ -21,6 +22,10 @@ dependencies {
     // No support for JUnit 6 in Mockito as of 2026-06. See https://github.com/mockito/mockito/pull/3781
     testImplementation libs.mockito.junit5
     testImplementation libs.log4j.core
+
+    // Important: Use the version of archunit-junit* that matches this package.
+    archTestImplementation project(path: ':archunit-junit6')
+    archTestRuntimeOnly libs.junit6PlatformLauncher
 }
 
 gradle.projectsEvaluated {
