+++ archunit/build.gradle
@@ -1,6 +1,7 @@
 plugins {
     id 'archunit.java-release-conventions'
     id 'archunit.java-junit-conventions'
+    id 'archunit.self-test-conventions'
 }
 
 ext.moduleName = 'com.tngtech.archunit'
@@ -20,6 +21,10 @@ dependencies {
         exclude module: 'assertj-core'
         exclude module: 'guava'
     }
+
+    // This subproject is JUnit-agnostic. We run the architecture self-tests using JUnit5 because of Java 8 compatibility.
+    archTestImplementation project(path: ':archunit-junit5')
+    archTestRuntimeOnly libs.junit5PlatformLauncher
 }
 
 shadowJar {
