+++ archunit/src/archTest/java/com/tngtech/archunit/ArchUnitArchitectureTest.java
@@ -1,0 +1,14 @@
+package com.tngtech.archunit;
+
+import com.tngtech.archunit.junit.AnalyzeClasses;
+import com.tngtech.archunit.junit.ArchTest;
+import com.tngtech.archunit.junit.ArchTests;
+
+import static com.tngtech.archunit.ArchitectureTestImportOptions.DoNotIncludeSelfTests;
+
+@AnalyzeClasses(packages = "com.tngtech.archunit", importOptions = DoNotIncludeSelfTests.class)
+class ArchUnitArchitectureTest {
+
+    @ArchTest
+    static final ArchTests architecture_rules = ArchTests.in(ArchUnitArchitectureRules.class);
+}
